@@ -1,5 +1,5 @@
-using GeneticOptimize;
 
+using GeneticOptimize;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using System;
@@ -9,16 +9,16 @@ namespace GeneticOptimizeTest {
     public class BoolCodonTest {
         [TestMethod]
         public void CreateTest() {
-            Random random = new Random();
+            Random random = new();
 
             var codon = new BoolCodon(random);
 
-            CollectionAssert.Contains(new bool[]{ true, false }, (bool)codon);
+            CollectionAssert.Contains(new bool[] { true, false }, (bool)codon);
         }
 
         [TestMethod]
         public void MutateTest() {
-            Random random = new Random();
+            Random random = new();
 
             var codon = new BoolCodon(random);
 
