@@ -1,5 +1,5 @@
-using GeneticOptimize;
 
+using GeneticOptimize;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using System;
@@ -9,26 +9,30 @@ namespace GeneticOptimizeTest {
     public class IntCodonTest {
         [TestMethod]
         public void CreateTest() {
-            Random random = new Random();
+            Random random = new();
 
-            var codon = new IntCodon(random, 5);
+            for (int i = 0; i < 100; i++) {
+                var codon = new IntCodon(random, 5);
 
-            Assert.IsTrue(2 <= codon && codon <= 5);
+                Assert.IsTrue(0 <= codon && codon < 5);
+            }
         }
 
         [TestMethod]
         public void MutateTest() {
-            Random random = new Random();
+            Random random = new();
 
             var codon = new IntCodon(random, 5);
 
             Console.WriteLine(codon);
 
-            codon.Mutate(random);
+            for (int i = 0; i < 100; i++) {
+                codon.Mutate(random);
 
-            Assert.IsTrue(2 <= codon && codon <= 5);
+                Assert.IsTrue(0 <= codon && codon < 5);
 
-            Console.WriteLine(codon);
+                Console.WriteLine(codon);
+            }
         }
     }
 }
