@@ -1,5 +1,5 @@
-using GeneticOptimize;
 
+using GeneticOptimize;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using System;
@@ -9,18 +9,18 @@ namespace GeneticOptimizeTest {
     public class BoolGeneTest {
         [TestMethod]
         public void CreateTest() {
-            Random random = new Random();
+            Random random = new();
 
-            BoolGene gene = new BoolGene(random, 10);
+            BoolGene gene = new(random, 10);
 
             Console.WriteLine(gene);
         }
 
         [TestMethod]
         public void MutateTest() {
-            Random random = new Random();
+            Random random = new();
 
-            BoolGene gene = new BoolGene(random, 10);
+            BoolGene gene = new(random, 10);
 
             Console.WriteLine(gene);
 
@@ -31,16 +31,16 @@ namespace GeneticOptimizeTest {
 
         [TestMethod]
         public void CrossoverTest() {
-            Random random = new Random();
+            Random random = new();
 
-            BoolGene gene1 = new BoolGene(random, 10);
-            BoolGene gene2 = new BoolGene(random, 10);
+            BoolGene gene1 = new(random, 10);
+            BoolGene gene2 = new(random, 10);
 
             Console.WriteLine(gene1);
             Console.WriteLine(gene2);
 
-            BoolGene gene_cross = new BoolGene(random, 10);
-            
+            BoolGene gene_cross = new(random, 10);
+
             gene_cross.Crossover(random, gene1, gene2);
 
             Console.WriteLine(gene_cross);
