@@ -1,5 +1,5 @@
-using GeneticOptimize;
 
+using GeneticOptimize;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using System;
@@ -9,18 +9,18 @@ namespace GeneticOptimizeTest {
     public class IntGeneTest {
         [TestMethod]
         public void CreateTest() {
-            Random random = new Random();
+            Random random = new();
 
-            IntGene gene = new IntGene(random, 10, 5);
+            IntGene gene = new(random, 10, 5);
 
             Console.WriteLine(gene);
         }
 
         [TestMethod]
         public void MutateTest() {
-            Random random = new Random();
+            Random random = new();
 
-            IntGene gene = new IntGene(random, 10, 5);
+            IntGene gene = new(random, 10, 5);
 
             Console.WriteLine(gene);
 
@@ -31,16 +31,16 @@ namespace GeneticOptimizeTest {
 
         [TestMethod]
         public void CrossoverTest() {
-            Random random = new Random();
+            Random random = new();
 
-            IntGene gene1 = new IntGene(random, 10, 5);
-            IntGene gene2 = new IntGene(random, 10, 5);
+            IntGene gene1 = new(random, 10, 5);
+            IntGene gene2 = new(random, 10, 5);
 
             Console.WriteLine(gene1);
             Console.WriteLine(gene2);
 
-            IntGene gene_cross = new IntGene(random, 10, 5);
-            
+            IntGene gene_cross = new(random, 10, 5);
+
             gene_cross.Crossover(random, gene1, gene2);
 
             Console.WriteLine(gene_cross);
