@@ -1,9 +1,8 @@
-using GeneticOptimize;
 
+using GeneticOptimize;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
 
 namespace GeneticOptimizeTest {
@@ -11,27 +10,27 @@ namespace GeneticOptimizeTest {
     public class RandomExtendTest {
         [TestMethod]
         public void BoolTest() {
-            Random random = new Random();
+            Random random = new();
 
-            for(int i = 0; i < 10; i++) { 
+            for (int i = 0; i < 10; i++) {
                 Console.WriteLine(random.NextBool());
             }
 
-            for(int i = 0; i < 10; i++) { 
+            for (int i = 0; i < 10; i++) {
                 Console.WriteLine(random.NextBool(0.2));
             }
         }
 
         [TestMethod]
         public void BoolsTest() {
-            Random random = new Random();
+            Random random = new();
 
             int trues = 0, falses = 0;
 
-            foreach(int length in new int[]{ 7, 8, 9, 15, 16, 17 }) { 
+            foreach (int length in new int[] { 7, 8, 9, 15, 16, 17 }) {
                 Console.WriteLine(length);
-                
-                foreach(bool b in random.NextBools(length)) { 
+
+                foreach (bool b in random.NextBools(length)) {
                     Console.WriteLine(b);
 
                     if (b) {
@@ -49,25 +48,25 @@ namespace GeneticOptimizeTest {
 
         [TestMethod]
         public void RangeTest() {
-            Random random = new Random();
+            Random random = new();
 
-            for(int i = 0; i < 10; i++) { 
+            for (int i = 0; i < 10; i++) {
                 Console.WriteLine(random.NextRange(0.1, 0.5));
             }
         }
 
         [TestMethod]
         public void ChoiceTest() {
-            Random random = new Random();
+            Random random = new();
 
             IEnumerable<int> xs = new int[] { 1, 2, 3, 4 };
 
-            for(int i = 0; i < 10; i++) { 
+            for (int i = 0; i < 10; i++) {
                 Console.WriteLine(random.Choice(xs));
             }
 
-            for(int i = 0; i < 10; i++) { 
-                Console.WriteLine(random.Choice( new int[] { 5, 6, 7, 8 }));
+            for (int i = 0; i < 10; i++) {
+                Console.WriteLine(random.Choice(new int[] { 5, 6, 7, 8 }));
             }
         }
     }
