@@ -1,5 +1,5 @@
-﻿using GeneticOptimize;
-
+﻿
+using GeneticOptimize;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 using System;
@@ -17,9 +17,9 @@ namespace GeneticOptimizeTest {
             int[] values = new int[items] { 10, 8, 6, 5, 6, 7, 4, 20, 5, 6, 3, 7, 8, 9, 2, 15, 20, 20, 2, 6 };
             int[] weights = new int[items] { 7, 6, 7, 4, 5, 3, 10, 5, 4, 4, 5, 9, 2, 5, 4, 7, 8, 9, 10, 22 };
 
-            Random random = new Random();
+            Random random = new();
 
-            NapsackOptimizer napsack = new NapsackOptimizer(items, values, weights, max_weight, genes:20, gene_saves:5, mutate_prob: 0.1);
+            NapsackOptimizer napsack = new(items, values, weights, max_weight, genes: 20, gene_saves: 5, mutate_prob: 0.1);
 
             napsack.Initialize(random);
 
