@@ -1,6 +1,6 @@
 ﻿using System;
-using System.Linq;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace GeneticOptimize {
     public abstract class Optimizer<Codon> where Codon : ICodon {
